backend:ucx: Implement handler reuse (#172)
* backend/ucx_mo: Fix error check

Fix a typo in the ucx_mo backend test that assumes that
prepXfer can return "in-progress"

Signed-off-by: Artem Y. Polyakov <artemp@nvidia.com>

* backend/ucx: Introduce handler reuse testing

Implement handler reuse testing code in UCX backend test

Signed-off-by: Artem Y. Polyakov <artemp@nvidia.com>

* backend/ucx: Implement handler reuse in UCX backend

Add invariant data structure that is not outdated after
the transfer is completed.
This allows to reuse the same handler multiple times.

Signed-off-by: Artem Y. Polyakov <artemp@nvidia.com>

* backend/ucx_mo: Update to support UCX prepXfer

As UCX backend moved to rely on prepXfer,
this needs to be reflected in UCX-MO backend too.

Signed-off-by: Artem Y. Polyakov <artemp@nvidia.com>

* core: Remove handle workaround that is not needed anymore

The code was added as a workaround for UCX backend not
allowing to reuse the handle.
Now as it supports the reuse, this workaround breaks things.

Signed-off-by: Artem Y. Polyakov <artemp@nvidia.com>

---------

Signed-off-by: Artem Y. Polyakov <artemp@nvidia.com>