Implement multi-threading support (#151)
* Implement multi-threading support

Current NIXL implementation sidesteps any considerations of executing in
multi-threaded environment by leaving it up to the user to ensure that
access to nixlAgent instance is synchronized. Extend this by adding a
strict synchronization mode while also preserving existing mode but making
it explicit.

Synchronization model is configured via nixlAgentConfig->syncMode field:

- When set to NIXL_THREAD_SYNC_NONE no locking is performed and behavior is
similar as before this change with the small difference that dynamic check
is performed every time the lock would be taken with strict model.

- When set to NIXL_THREAD_SYNC_STRICT all nixlAgent API methods are fully
synchronized with internal exclusive lock.

In order to allow dynamic configuration at runtime the lock implemented as
a thin BasicLockable-compatible wrapper around std::mutex that accepts
sync_mode parameter to its constructor and skip locking/unlocking the mutex
when synchronization mode is not strict.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

* Refactor nixl plugin handle for const-friendliness

Current implementation of nixlPluginHandle class already doesn't change its
internal state after construction. To facilitate refactoring
nixlPluginManager to concurrency-safe version in the following patches make
assumptions about nixlPluginHandle immutability with following changes:

- Add "const" modifier to all plugin handle methods.

- Return shared_ptr to const plugin handle instance from all manager
function that were returning a mutable shared_ptr before the change.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

* Refactor static plugins handling

The current implementation splits maintenance of static static_plugin_
vector between static and dynamic methods of nixlPluginManager, e.g. the
vector is filled by dynamic registerBuiltinPlugins() and
registerStaticPlugin() methods but is itself declared as static and its
accessor getStaticPlugins() is also static.

Considering that such approach doesn't make much sense and that
nixlPluginManager is already a singleton refactor static plugins to be
dynamically managed by the singleton instance.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

* Make nixlPluginManager initialization thread-safe

Currently the initialization of nixlPluginManager class is split between
constructor and registerBuiltinPlugins() member function that is called
only once in getInstance(). This doesn't seem to be caused by any
necessity, so just call the init function from ctor and remove all the
machinery required to call it only once in a multi-threaded environment.

The call to ctor of static function-local instance of nixlPluginManager is
already thread safe in C++11 and later (Meyers singleton).

While we are at it, also make registerStaticPlugin() method private since
it is only called from registerStaticPlugin which is now just a constructor
helper and remove the locking from it.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

* Lock in plugin manager methods that modify its state

After preparatory changes in previous commits that made plugin handle usage
const and refactored nixlPluginManager singleton initialization for thread
safety finish the job by extending nixlPluginManager with a mutex lock and
using the lock to synchronize parts of the code that modify or read shared
mutable state.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

* Implement multi-threading tests

New gtest-based infrastructure has recently been added. Extend it for
validating multi-threaded execution of the library:

 - Modify MockDramBackendEngine to have a shared mutable state represented
 by an integer sharedState. Read the integer value with assert in all const
 mock methods and increment its value in non-const ones.

 - Implement new 'MultiThreadedTestFixture' test that executes several
 typical use-cases with the mock engine (independent nixlAgent object
 per-thread, shared nixlAgent object, etc.) on multiple threads.

The test is only executed when the project is compiled with Thread
sanitizer (-Db_sanitize=thread) in order to catch any concurrency-unsafe
behavior.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

---------

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>