Refactor UCX progress thread to remove busy-polling (#297)
* Refactor UCX progress thread to remove busy-polling

Instead of busy-polling use functionality provided by UCX that exposes file
descriptors that can be monitored for events with standard UNIX APIs.
Implement following changes :

 - When progress thread is enabled also enable UCP feature that exposes
 file descriptors for wake up.

 - In UCX progress thread arm every worker for wake up via file descriptor
 and sleep while waiting for events with poll().

 - Use control pipe file descriptor to wake up the thread for cancellation.

 - Fix uninitialized variable usage in tests
On error cudaGetDeviceCount() doesn't set its input pointer argument value
and it is used afterwards with value of whatever garbage happened to be on
that stack address. Fix it by setting the variable to zero on error.

---------

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>