contrib: Remove the lib.real from the LD_LIBRARY path (#466)
* contrib: Remove the lib.real from the LD_LIBRARY path

Keeping lib.real in the env var LD_LIBRARY_PATH makes a mess for the
CUDA scripts to setup the symlinks for the nvidia runtime. This can
cause the error 803 to be displayed.

Unfortunately, we need the libcuda.so.1 when repairing the wheel, so
move the ENV command there and restore the old LD_LIBRARY_PATH after
finishing.

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>

* benchmark: Remove lib.real and wheels creation

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>

* contrib: Remove lib.real from manylinux Dockerfile

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>

* benchmark: Keep nixlbench Dockerfile same as others

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>

* contrib: Move the LD_LIBRARY_PATH to original locations

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>

---------

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>