Fix nixlbench: properly implement --num_files, fix READ mode support (#635)
* nixlbench: handle short writes

Some file systems are capping writes at 2GB.

Signed-off-by: Dan Aloni <dan.aloni@vastdata.com>

* Fix nixlbench: properly implement --num_files, fix READ mode support

So far, giving `--num_files <n>` with a value of 1 resulted in one file
with the total buffer size, while the other files ended up with file
size 0.

- Allow dividing M buffers to N files, checking that M is divisible by N,
  and that each file contains at least one buffer.
- READ mode does not need to write to the files if they are big enough.
- `buffer_size` is already made sure to be page aligned for direct IO in `allocateMemory`,
  no need to do the alignment again in `initBasicDescFile`.
- No need to recall `initBasicDescFile` in `exchangeIOV`, as we don't
  want to do writes to the file for the second time. We only need to
  change `len` from the `remote_iovs` that were already computed. The
  offset in the file will remain the same, only len will be smaller than
  the maximum.
- Remove `gds_running_ptr` global. Allocate a `buffer_size` from each file,
  so we need to track what offsets we already used using the newly added
  xferFileState.
- Remove unused globals gds_remote_iovs, storage_remote_iovs, gds_remote_iovs

Signed-off-by: Dan Aloni <dan.aloni@vastdata.com>

* Fix nixlbench: batch size greater than 1, proper file offsets

The previous change done here, should have retained the creation of remote_iovs
from local_iovs so that the amount of bytes to transfer in both lists is
identical.

We just need to make sure the remote IOV has the FD of the file, a file
offset and proper size. Before the PR, all batches wrote to file offset 0, which
was a bug, because the file size was calculated to allow for all buffers having
their own offset.

Signed-off-by: Dan Aloni <dan.aloni@vastdata.com>

---------

Signed-off-by: Dan Aloni <dan.aloni@vastdata.com>
Co-authored-by: Adit Ranadive <aranadive@nvidia.com>
Co-authored-by: ovidiusm <ovidium@nvidia.com>